change structure of files to allow for individual modules
